Filter on both municipality and zone use
The filtering is still very slow, so it helps to be zoomed in on your area of interest. Further, the default WHERE should show the entire map, whereas it currently shows nothing. The next step is to implement conditionality to each filter category and not include it if there are no filterValues