Wire up single-layer fix to project correct data
While we work on the simplified full-view layer, we now pull correct data from the full zoning layer. The result of this is a slightly slower-loading map, but with the expected data. Also un-commenting the "Municipality" section of the tooltip because the filtering issue is resolved on the full layer.